+++ tests/test_startat.py
@@ -1,21 +1,20 @@
 #!/usr/bin/env python3
-"""Test the location of the csv file"""
-# pylint: disable=duplicate-code
+"""Test `trk start --at`"""
 
 from os.path import exists
-from logging import basicConfig
-from logging import DEBUG
+#from logging import basicConfig
+#from logging import DEBUG
 from logging import debug
 from tempfile import TemporaryDirectory
 from timetracker.utils import cyan
-from timetracker.cfg.finder import CfgFinder
+from timetracker.utils import yellow
 from timetracker.cmd.init import run_init_test
 from timetracker.cmd.start import run_start
-from tests.pkgtttest.mkprojs import mk_projdirs
-from tests.pkgtttest.mkprojs import findhome_str
-from tests.pkgtttest.startdts import DT2525
+from tests.pkgtttest.dts import DT2525
+from tests.pkgtttest.runfncs import RunBase
+from tests.pkgtttest.runfncs import proj_setup
 
-basicConfig(level=DEBUG)
+#basicConfig(level=DEBUG)
 
 SEP = f'\n{"="*80}\n'
 
@@ -26,6 +25,7 @@ def test_startat(project='pumpkin', username='carver'):
     _run(Obj(project, username, dircur='dirdoc',  dirgit01=True))
 
 def _run(obj):
+    # pylint: disable=duplicate-code
     # Test researcher-entered datetime starttimes
     obj.chk('4am',                   '2525-01-01 04:00:00')
     obj.chk("2025-02-19 17:00:00",   '2025-02-19 17:00:00')
@@ -46,25 +46,16 @@ def _run(obj):
     obj.chk("4:00:00",    '2525-01-01 04:00:00')
 
 
-class Obj:
-    """Test the location of the csv file"""
+class Obj(RunBase):
+    """Test `trk start --at`"""
     # pylint: disable=too-few-public-methods
-
-    def __init__(self, project, username, dircur, dirgit01):
-        self.project = project
-        self.uname = username
-        self.dircurattr = dircur
-        self.dirgit01 = dirgit01
 
     def _run(self, start_at, dircsv=None):
         """Run init, start --at, stop"""
         debug(cyan(f'\n{"="*100}'))
         debug(cyan(f'RUN(start_at={start_at})'))
         with TemporaryDirectory() as tmphome:
-            exp = mk_projdirs(tmphome, self.project, self.dirgit01)
-            finder = CfgFinder(dircur=getattr(exp, self.dircurattr), trksubdir=None)
-            cfgname = finder.get_cfgfilename()
-            assert not exists(cfgname), findhome_str(exp.dirhome)
+            cfgname, _, exp = proj_setup(tmphome, self.project, self.dircur, self.dirgit01)
 
             # CMD: INIT; CFG PROJECT
             cfgp, _ = run_init_test(cfgname, dircsv, self.project, exp.dirhome)  # cfgg
@@ -72,17 +63,16 @@ class Obj:
 
             # CMD: START
             fin_start = run_start(cfgname, self.uname,
-                                  start_at=start_at,
-                                  now=DT2525,
-                                  defaultdt=DT2525)
+                                  start_at=start_at, now=DT2525, defaultdt=DT2525)
             assert exists(fin_start)
             return cfgp.get_starttime_obj(self.uname).read_starttime()
 
 
     def chk(self, start_at, expstr):
         """Run start --at and check value"""
-        starttime = self._run(start_at)
-        assert str(starttime) == expstr, f'TEST({start_at}): ACT({starttime}) !=  EXP({expstr})'
+        print(yellow(f'\nTEST: start={start_at:22} EXP={expstr}'))
+        startdt = self._run(start_at)
+        assert str(startdt) == expstr, f'TEST({start_at}): ACT({startdt}) !=  EXP({expstr})'
 
 
 if __name__ == '__main__':