+++ timetracker/cfg/cfg_local.py
@@ -19,6 +19,7 @@ from os.path import abspath
 from os.path import dirname
 from os.path import normpath
 from logging import debug
+from glob import glob
 
 from tomlkit import comment
 from tomlkit import document
@@ -69,9 +70,19 @@ class CfgProj:
         """Get the csv filename by reading the cfg csv pattern and filling in"""
         username = get_username(username)
         fcsv = self._read_csv_from_cfgfile(username)
-        if fcsv is not None:
-            return fcsv
-        return replace_envvar(self._get_dircsv_absname(), username)
+        ####if fcsv is not None:
+        ####    return fcsv
+        ####return replace_envvar(self._get_dircsv_absname(), username)
+        ####return None
+        return fcsv if fcsv is not None else None
+
+    def get_project_csvs(self):
+        """Get the csv filename by reading the cfg csv pattern and filling in"""
+        fcsvpat = self._read_csvpat_from_cfgfile()
+        if fcsvpat is not None:
+            globpat = replace_envvar(fcsvpat, '*')
+            return glob(globpat)
+        return None
 
     def set_filename_csv(self, filename_str):
         """Write the config file, replacing [csv][filename] value"""
@@ -103,9 +114,11 @@ class CfgProj:
             doc = self._get_doc_new()
             doc['csv']['filename'] = join(dircsv, self.CSVPAT)
             self._wr_cfg(fname, doc)
-            print(f'Overwrote {fname}')
+            if not quiet:
+                print(f'Overwrote {fname}')
         else:
-            print(f'Use `force` to overwrite: {fname}')
+            if not quiet:
+                print(f'Use `force` to overwrite: {fname}')
 
     def read_doc(self):
         """Read a config file and load it into a TOML document"""
@@ -132,11 +145,22 @@ class CfgProj:
 
     def _read_csv_from_cfgfile(self, username):
         """Read a config file and load it into a TOML document"""
+        ####doc = self.read_doc()
+        ####if doc is not None:
+        ####    fpat = get_abspath(doc['csv']['filename'], self.dirproj, self.dirhome)
+        ####    fpat = fpat.replace('PROJECT', self.project)
+        fcsvpat = self._read_csvpat_from_cfgfile()
+        if fcsvpat:
+            return replace_envvar(fcsvpat, username) if '$' in fcsvpat else fcsvpat
+        return None
+
+    def _read_csvpat_from_cfgfile(self):
+        """Read a config file and load it into a TOML document"""
         doc = self.read_doc()
         if doc is not None:
             fpat = get_abspath(doc['csv']['filename'], self.dirproj, self.dirhome)
             fpat = fpat.replace('PROJECT', self.project)
-            return replace_envvar(fpat, username) if '$' in fpat else fpat
+            return fpat
         return None
 
     def _read_csvdir_from_cfgfile(self):