+++ timetracker/timecalc.py
@@ -19,9 +19,33 @@ from datetime import timedelta
 
 def timedelta_to_hms(tdelta):
     """Convert a timedelta to hours and minutes"""
-    hours, remainder = divmod(tdelta.total_seconds(), 3600)
+    total_secs = tdelta.total_seconds()
+    assert total_secs >= 0, total_secs
+    hours, remainder = divmod(total_secs, 3600)
     minutes, seconds = divmod(remainder, 60)
     return int(hours), int(minutes), seconds
+
+def _totsecs_to_hms(total_secs):
+    """Convert a timedelta to hours and minutes"""
+    hours, remainder = divmod(total_secs, 3600)
+    minutes, seconds = divmod(remainder, 60)
+    return int(hours), int(minutes), seconds
+
+def str_td(tdelta):
+    """Convert a tuple containing hours, minutes, and seconds to a string"""
+    total_secs = tdelta.total_seconds()
+    return _secspos_to_str(total_secs) if total_secs >= 0 else _secsneg_to_str(total_secs)
+
+def _secspos_to_str(total_secs):
+    """Convert a tuple containing hours, minutes, and seconds to a string"""
+    cur_hours, cur_minutes, _ = _totsecs_to_hms(total_secs)
+    return f'{cur_hours:02}:{cur_minutes:02}'
+
+def _secsneg_to_str(total_secs):
+    """Convert a tuple containing hours, minutes, and seconds to a string"""
+    cur_hours, cur_minutes, _ = _totsecs_to_hms(-total_secs)
+    return f'-{cur_hours:02}:{cur_minutes:02}'
+
 
 class RoundTime:
     """Round a datetime object up or down to `round_to_min`"""